+++ sb.py
@@ -61,7 +61,7 @@ def create_project(spec, name, labels, logger, **kwargs):
     for secret in resp.secrets:
         if f'{name}-token' in secret.name:
             secret_response = core_v1.read_namespaced_secret(namespace=name, name=secret.name)
-            response = requests.post(f"{sb_url}/projects/{project_uuid}/token/", json=secret_response.data)
+            response = requests.post(f"{sb_url}/projects/{project_uuid}/token/", json=secret_response.data, verify=False)
             logger.info(f"Sent service account token for project {name}.")
 
 
@@ -118,11 +118,15 @@ def create_app(spec, name, labels, namespace, logger, **kwargs):
     logger.debug(f"An application is created with spec: {spec}")
     api = client.CustomObjectsApi()
     # create service account
-    service_account = create_service_account(name, namespace, logger)
-
+    #TODO: handle exception if already created
+    try:
+        service_account = create_service_account(name, namespace, logger)
+    except:
+        logger.debug("Serivce account already exists.")
     # add secret to service account if private repo
     git_info = spec.get('git')
     repo = git_info.get('repo')
+    # TODO: This is not a hard enough check.
     if repo.startswith('git@'):
         logger.info('Attaching ssh secret.')
         core_v1 = client.CoreV1Api()
@@ -136,6 +140,8 @@ def create_app(spec, name, labels, namespace, logger, **kwargs):
     response = requests.get(f"{sb_url}/apps/{app_uuid}/last-deployment/")
     if response.status_code == 200:
         deployment_uuid = response.json()['deployment']
+    else:
+        deployment_uuid = None
     # Create builder
     try:
         resource = api.get_namespaced_custom_object(
@@ -167,10 +173,11 @@ def create_app(spec, name, labels, namespace, logger, **kwargs):
             'logs': 'Builder created.\n',
             'status': 'running',
             'app_uuid': app_uuid,
-            'deployment_uuid': deployment_uuid,
             }
+            if deployment_uuid:
+                data['deployment_uuid'] = deployment_uuid
             pusher_client.trigger(str(app_uuid), 'deployment', data)
-            response = requests.post(f"{sb_url}/deployments/", json=data)
+            response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
             #response = requests.post(f"{sb_url}/helm-status/", json=data)
 
     # create image
@@ -215,7 +222,7 @@ def create_app(spec, name, labels, namespace, logger, **kwargs):
             'deployment_uuid': deployment_uuid,
             }
             pusher_client.trigger(str(app_uuid), 'deployment', data)
-            response = requests.post(f"{sb_url}/deployments/", json=data)
+            response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
     return {'lastDeployment': deployment_uuid}
 
 @kopf.on.update('kpack.io', 'v1alpha2', 'builds')
@@ -244,7 +251,7 @@ def update_build(spec, status, name, namespace, logger, labels, **kwargs):
         else:
             if len(data['logs']):
                 pusher_client.trigger(str(app_uuid), 'deployment', data)
-        response = requests.post(f"{sb_url}/deployments/", json=data)
+        response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
 
 @kopf.on.field('kpack.io', 'v1alpha2', 'builds', field='status.conditions')
 def trigger_helm_release(name, namespace, labels, spec, status, new, logger, **kwargs):
@@ -271,7 +278,7 @@ def trigger_helm_release(name, namespace, labels, spec, status, new, logger, **k
             'deployment_uuid': deployment_uuid,
         }
         pusher_client.trigger(str(app_uuid), 'deployment', data)
-        response = requests.post(f"{sb_url}/deployments/", json=data)
+        response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
         create_helmrelease(app_name, app_uuid, namespace, tag, logger)
         # Update the latest image tag in app status
         update_app_status(namespace, app_name, tag, logger)
@@ -283,7 +290,7 @@ def trigger_helm_release(name, namespace, labels, spec, status, new, logger, **k
             'deployment_uuid': deployment_uuid,
         }
         pusher_client.trigger(str(app_uuid), 'deployment', data)
-        response = requests.post(f"{sb_url}/deployments/", json=data)
+        response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
 
 def get_last_tag(status: Dict):
     """
@@ -352,7 +359,7 @@ def update_app(spec, name, namespace, logger, labels, status, **kwargs):
         'deployment_uuid': deployment_uuid,
     }
     pusher_client.trigger(str(app_uuid), 'deployment', data)
-    response = requests.post(f"{sb_url}/deployments/", json=data)
+    response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
     return {'lastDeployment': deployment_uuid}
 
 
@@ -376,7 +383,7 @@ def notify_helm_release(old, new, labels, diff, namespace, name, logger, **kwarg
     }
     pusher_client.trigger(str(app_uuid), 'deployment', data)
     logger.info(f"Updated Helm release status for {name} successfully.")
-    response = requests.post(f"{sb_url}/deployments/", json=data)
+    response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
 
 
 @kopf.on.delete('applications')
@@ -463,11 +470,11 @@ def send_cluster_admin_account(logger):
         'ca.crt': open('/var/run/secrets/kubernetes.io/serviceaccount/ca.crt').read(),
     }
     logger.info('POSTing token info.')
-    response = requests.post(f"{sb_url}/clusters/{cluster_id}/token-info", json=data)
+    response = requests.post(f"{sb_url}/clusters/{cluster_id}/token-info", json=data, verify=False)
     if response.status_code == 202:
         logger.info("POSTed token info.")
     nodes = get_nodes_info()
-    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes", json=nodes)
+    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes", json=nodes, verify=False)
     logger.info("POSTing node info")
     logger.info(response.status_code)
     if response.status_code == 201:
@@ -478,7 +485,7 @@ def send_cluster_admin_account(logger):
 @kopf.on.cleanup()
 async def cleanup_fn(logger, **kwargs):
     nodes = get_nodes_info()
-    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes-delete", json=nodes)
+    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes-delete", json=nodes, verify=False)
     if response.status_code == 201:
         logger.info("DELETE node info.")
 
@@ -548,7 +555,7 @@ def create_helmrelease(name, app_uuid, namespace, tag, logger):
             'deployment_uuid': deployment_uuid,
         }
         pusher_client.trigger(str(app_uuid), 'deployment', data)
-        response = requests.post(f"{sb_url}/deployments/", json=data)
+        response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
     except ApiException as error:
         if error.status == 404:
             chart_info = spec.get('chart')
@@ -586,7 +593,7 @@ def create_helmrelease(name, app_uuid, namespace, tag, logger):
             'deployment_uuid': deployment_uuid,
             }
             pusher_client.trigger(str(app_uuid), 'deployment', data)
-            response = requests.post(f"{sb_url}/deployments/", json=data)
+            response = requests.post(f"{sb_url}/deployments/", json=data, verify=False)
             #response = requests.post(f"{sb_url}/helm-status/", json=data)
 
 
@@ -676,7 +683,7 @@ def add_node(status, name, logger, **kwargs):
         'memory': status['capacity']['memory'],
     }
     node_data.append(node_info)
-    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes", json=node_data)
+    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes", json=node_data, verify=False)
     if response.status_code == 201:
         logger.info(f"POSTed node info for node {name}.")
 
@@ -688,7 +695,7 @@ def remove_node(status, name, logger, **kwargs):
         'memory': status['capacity']['memory'],
     }
     node_data.append(node_info)
-    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes-delete", json=node_data)
+    response = requests.post(f"{sb_url}/clusters/{cluster_id}/nodes-delete", json=node_data, verify=False)
     if response.status_code == 201:
         logger.info(f"DELETEd node info for node {name}.")
 
@@ -707,9 +714,10 @@ def get_nodes_info():
 
 # Update the stack run image every 12 hours
 # TODO: update the build image as well
+"""
 @kopf.timer('kpack.io', 'v1alpha2', 'clusterstack', interval=(3600.0 * 12))
 def update_run_image_sha(name, spec, status, logger, **kwargs):
-    logger.info(status['runImage']['latestImage'])
+    logger.info(status['runImage'].get('latestImage'))
     # get spec run image, if having not having sha, update.
     run_image = spec['runImage']['image']
     api = client.CustomObjectsApi()
@@ -747,3 +755,4 @@ def update_run_image(api, logger, run_image_sha):
         body=patch_body,
     )
     logger.info("Clusterstack patched.")
+"""